+++ src/entity/Entity.java
@@ -5,8 +5,8 @@ import java.awt.image.BufferedImage;
 
 public class Entity {
 
-    public double worldX, worldY;
-    public int speed;
+    public float worldX, worldY;
+    public float speed;
 
     public BufferedImage down1, down2, down3, down4, up1, up2, up3, up4, right1, right2, right3, right4, left1, left2, left3, left4;
     public String direction;
